+++ include/vnx/AsyncClient.h
@@ -58,6 +58,12 @@ public:
 	/// Sets a custom session to be used for requests.
 	void vnx_set_session(Hash64 id);
 
+	/// Sets a tunnel address for sending requests to.
+	void vnx_set_tunnel(Hash64 tunnel_addr);
+	
+	/// Sets a tunnel name for sending requests to.
+	void vnx_set_tunnel(const std::string& tunnel_name);
+
 	/// If \p non_blocking_mode == true requests will throw an exception instead of blocking (default = false)
 	void vnx_set_non_blocking(bool non_blocking_mode);
 
@@ -95,6 +101,7 @@ protected:
 private:
 	Node* vnx_node = nullptr;
 	Hash64 vnx_src_mac;
+	Hash64 vnx_tunnel_addr;
 	Hash64 vnx_service_addr;
 	Hash64 vnx_session_id;
 	std::atomic_bool vnx_is_non_blocking {false};