+++ include/vnx/Stream.h
@@ -25,6 +25,8 @@ public:
 	Stream(const Stream& other) = delete;
 	Stream& operator=(const Stream& other) = delete;
 	
+	bool open();
+	
 	bool connect(	Node* node,
 					int max_queue_ms, int max_queue_size = 0,
 					int priority = PIPE_PRIORITY_DEFAULT);