+++ include/vnx/ToStringValue.h
@@ -25,9 +25,8 @@ namespace vnx {
 /** \brief A Visitor that generates JSON string output with some exceptions.
  *
  * Same as DefaultPrinter except:
- *  - null is an empty string (instead of "{}")
- *  - strings are without quotes
- *  - enum values are without quotes
+ *  - strings are without quotes (at the top level)
+ *  - enum values are without quotes (at the top level)
  *
  * This Visitor is used when the desired output is a value.
  */
@@ -35,11 +34,7 @@ class ToStringValue : public DefaultPrinter {
 public:
 	ToStringValue(std::ostream& out);
 
-	void visit_null() override;
-
 	void visit(const std::string& value) override;
-
-	void enum_value(uint32_t value, const std::string& name) override;
 
 };